+++ src/lib/regexpcreator.js
@@ -54,6 +54,8 @@ class RegExpCreator {
       'accuracy': 'partially',
       'caseSensitive': false,
       'ignoreJoiners': false,
+      'characterSets': false,
+      'unicode': false,
       'ignorePunctuation': [],
       'wildcards': 'disabled'
     }, options);
@@ -83,33 +85,33 @@ class RegExpCreator {
     * @return {RegExp}
     */
   create(terms) {
-    const array = [];
-    let index = 0;
+    const array = [],
+      charSets = this.opt.characterSets;
+    let index = 0,
+      flags = `g${this.opt.caseSensitive ? '' : 'i'}${this.opt.unicode ? 'u' : ''}`;
 
     terms = terms.map(str => {
-      if (this.opt.charSets) {
+      if (charSets) {
         // saves RegExp character sets in the array and creates index placeholders to restore them later.
         // matches any escaped char | character set with quantifier
         str = str.replace(/(\\.)+|\[(?:[^\\\]]|(?:\\.))+\](?:[+*?]\??|\{[\d,]+\}\??)?/g, (m, gr) => {
           if (gr) return m;
-
           array.push(m);
           return '\x03' + index++ + '\x03';
         }).replace(/\\(?=\[|\x03)/g, ''); // removes one backslash character before '[' and '\x03'
       }
       // wraps an individual term pattern in a capturing group; that allows to determine
       // in filter callback which term is currently matched
-      return '(' + this.createPattern(str) + ')';
+      return '(' + this.createPattern(str, flags) + ')';
     });
 
     const obj = this.createAccuracy(terms.join('|'));
-
+    // restores RegExp character sets
     if (array.length) {
-      // restores RegExp character sets
       obj.pattern = obj.pattern.replace(/\x03(\d+)\x03/g, (m, gr) => array[gr]);
     }
 
-    return new RegExp(`${obj.lookbehind}(${obj.pattern})${obj.lookahead}`, `g${this.opt.caseSensitive ? '' : 'i'}`);
+    return new RegExp(`${obj.lookbehind}(${obj.pattern})${obj.lookahead}`, flags);
   }
 
   /**
@@ -117,16 +119,13 @@ class RegExpCreator {
    * @param  {string} str - The search term to be used
    * @return {string}
    */
-  createPattern(str) {
+  createPattern(str, flags) {
     const wildcards = this.opt.wildcards !== 'disabled';
-    // normalizes white spaces
-    str = str.replace(/\s+/g, ' ');
+    str = this.checkWildcardsEscape(str);
 
-    if (wildcards) {
-      str = this.createPlaceholders(str);
+    if (this.opt.synonyms) {
+      str = this.createSynonyms(str, flags);
     }
-    // escapes RegExp special characters
-    str = str.replace(/[[\]/{}()*+?.\\^$|]/g, '\\$&');
 
     const joiners = this.getJoinersPunctuation();
     if (joiners) {
@@ -136,7 +135,8 @@ class RegExpCreator {
     if (this.opt.diacritics) {
       str = this.createDiacritics(str);
     }
-    str = str.replace(/ /g, '\\s+');
+
+    str = str.replace(/\s+/g, '[\\s]+');
 
     if (joiners) {
       str = str.split(/\x00+/).join(`[${joiners}]*`);
@@ -149,13 +149,27 @@ class RegExpCreator {
   }
 
   /**
+   * Escapes RegExp special characters
+   * @param {string} str - The string to escape
+   * @return {string}
+   */
+  escape(str) {
+    return str.replace(/[[\]/{}()*+?.\\^$|]/g, '\\$&');
+  }
+
+  /**
    * Splits string if val is string, removes duplicates, escape '-^]\\' which are special in RegExp characters set
+   * preserves unicode class '\p{..}', '\P{..}' and character escape '\u{..}' if option 'unicode' is present
    * @param {array|string} val - The parameter to process
    * @return {string}
    */
   preprocess(val) {
     if (val && val.length) {
-      return this.distinct(typeof val === 'string' ? val.split('') : val).join('').replace(/[-^\]\\]/g, '\\$&');
+      if (typeof val === 'string') {
+        val = this.opt.unicode ? val.split(/(\\[pPu]\{[^}]+\}|.)/) : val.split('');
+      }
+      // minimal length of unicode escape class, e.g. '\p{L}' is 5;
+      return this.distinct(val).map(ch => ch.length > 4 ? ch : ch.replace(/[-^\]\\]/g, '\\$&')).join('');
     }
     return '';
   }
@@ -168,7 +182,7 @@ class RegExpCreator {
   distinct(array) {
     const result = [];
     array.forEach(item => {
-      if (item.trim() && result.indexOf(item) === -1) {
+      if (item.trim() && !result.includes(item)) {
         result.push(item);
       }
     });
@@ -204,15 +218,45 @@ class RegExpCreator {
   }
 
   /**
+   * Creates a regular expression string to match the defined synonyms
+   * @param  {string} str - The search term to be used
+   * @return {string}
+   */
+  createSynonyms(str, flags) {
+    const syn = this.opt.synonyms;
+
+    for (const key in syn) {
+      if (syn.hasOwnProperty(key)) {
+        let array = Array.isArray(syn[key]) ? syn[key] : [syn[key]];
+        array.unshift(key);
+        array = this.distinct(array);
+
+        if (array.length > 1) {
+          array.sort((a, b) => a.length === b.length ? (a > b ? 1 : -1) : b.length - a.length);
+          array = array.map(term => this.checkWildcardsEscape(term));
+
+          const pattern = array.map(term => this.escape(term)).join('|');
+          str = str.replace(new RegExp(pattern, flags), array.join('|'));
+        }
+      }
+    }
+    return str;
+  }
+
+  /**
    * Creates placeholders in the regular expression string to allow later insertion of wildcard patterns.
    * Correctly handles escaping
    * @param {string} str - The search term
    * @return {string}
    */
-  createPlaceholders(str) {
-    return str.replace(/(\\.)+|[?*]/g, (m, gr) => gr ? m : m === '?' ? '\x01' : '\x02')
-      // removes one backslash character before '?', '*', '\x01', and '\x02'
-      .replace(/\\+(?=[?*\x01\x02])/g, m => m.slice(1));
+  checkWildcardsEscape(str) {
+    if (this.opt.wildcards !== 'disabled') {
+      // replaces single character wildcard with \x01, multiple character wildcard with \x02
+      str = str.replace(/(\\.)+|[?*]/g, (m, gr) => gr ? m : m === '?' ? '\x01' : '\x02')
+        // removes one backslash character before '?', '*', '\x01', and '\x02'
+        .replace(/\\+(?=[?*\x01\x02])/g, m => m.slice(1));
+    }
+    return this.escape(str);
   }
 
   /**
@@ -222,7 +266,7 @@ class RegExpCreator {
    */
   createWildcards(str) {
     const spaces = this.opt.wildcards === 'withSpaces',
-      anyChar = spaces && this.opt.boundary ? '[^\x01]*?' : '[^]*?';
+      anyChar = spaces && this.opt.blockElementsBoundary ? '[^\x01]*?' : '[^]*?';
 
     return str.replace(/\x01/g, spaces ? '[^]?' : '\\S?').replace(/\x02/g, spaces ? anyChar : '\\S*');
   }
@@ -261,9 +305,10 @@ class RegExpCreator {
    * @return {RegExpCreator~patternObj}
    */
   createAccuracy(str) {
+    const chars = '!-/:-@[-`{-~¡¿'; // '!"#$%&\'()*+,\\-./:;<=>?@[\\]\\\\^_`{|}~¡¿';
     let accuracy = this.opt.accuracy,
       lookbehind = '()',
-      pattern = str,
+      pattern = `(?:${str})`,
       lookahead = '',
       limiters;
 
@@ -273,22 +318,21 @@ class RegExpCreator {
         accuracy = accuracy.value;
       }
 
-      limiters = '\\s' + (limiters || '!"#$%&\'()*+,\\-./:;<=>?@[\\]\\\\^_`{|}~¡¿');
-      const group = `(^|[${limiters}])`;
-
       if (accuracy === 'exactly') {
-        lookbehind = group;
-        lookahead = `(?=$|[${limiters}])`;
+        const charSet = limiters ? '[\\s' + limiters + ']' : '\\s';
+        lookbehind = `(^|${charSet})`;
+        lookahead = `(?=$|${charSet})`;
 
       } else {
-        const charSet = `[^${limiters}]*`;
+        const chs = limiters || chars,
+          charSet = `[^\\s${chs}]*`;
 
         if (accuracy === 'complementary') {
-          pattern = `${charSet}(?:${str})${charSet}`;
+          pattern = charSet + pattern + charSet;
 
         } else if (accuracy === 'startsWith') {
-          lookbehind = group;
-          pattern = `(?:${str.replace(/\\s\+/g, charSet + '$&')})${charSet}`;
+          lookbehind = `(^|[\\s${chs}])`;
+          pattern = pattern.split(/\[\\s\]\+/).join(charSet + '[\\s]+') + charSet;
         }
       }
     }