+++ src/types/mark.d.ts
@@ -40,8 +40,7 @@ declare namespace Mark {
     iframes?: boolean;
     iframesTimeout?: number;
 
-    combinePatterns?: boolean;
-    cacheTextNodes?: boolean;
+    combineby?: number;
     blockElementsBoundary?: boolean | BoundaryObject;
     shadowDOM?: boolean | ShadowObject;
 
@@ -51,7 +50,7 @@ declare namespace Mark {
     each?(element: Element, eachInfo: MarkEachInfo) : void;
     done?(totalMarks: number, totalMatches: number, termStats: TermStats) : void;
 
-    noMatch?(term: string | string[]) : void;
+    noMatch?(term: string[]) : void;
     debug?: boolean;
     log?: object;
   }
@@ -60,7 +59,6 @@ declare namespace Mark {
     match: RegExpExecArray;
     matchStart: boolean;
     execution: ExecutionObject;
-    offset: number;
   }
 
   interface MarkEachInfo {
@@ -101,7 +99,6 @@ declare namespace Mark {
     matchStart: boolean;
     execution: ExecutionObject;
     groupIndex?: number;
-    offset?: number;
   }
 
   interface RegExpEachInfo {