+++ test/specs/across-elements/basic/accuracy-exactly.js
@@ -12,9 +12,9 @@ describe('mark with acrossElements and accuracy exactly', () => {
       'accuracy': 'exactly',
       'acrossElements': true,
       'done': () => {
-        expect($ctx1.find('mark').length).toBe(2);
+        expect($ctx1.find('mark').length).toBe(1);
         expect($ctx1.find('mark').first().text()).toBe('ipsu');
-        //expect($ctx1.find('.not mark').length).toBe(0);
+        expect($ctx1.find('.not mark').length).toBe(0);
         
         done();
       }